turbobujias-ai/requirements.txt
gradio==4.44.1
transformers>=4.39.0
sentence-transformers>=2.6.1
faiss-cpu>=1.7.4
langchain>=0.1.16
langchain-community>=0.0.36
huggingface-hub>=0.22.2,<1.0
openai-whisper>=20231117
torch>=2.2.2
accelerate>=0.29.2
python-dotenv>=1.0.1
requests>=2.32.0
